Modal Issues (description included)
Modal is working for edit items, however only the first edit button can open it.

Also when on the manage lists page the login button opens the edit modal. Not sure how much of an issue that is right now but it could be an issue when using that area for logout.

Might want to think about a dropdown style after logged in. Could have edit account info and logout option.